--- Playgrounds/Chat Basics/MessagesCollectionLayoutExample/Messages View Controller/Controller/Image Loader/DefaultImageLoader.swift
@@ -1,55 +1,0 @@
-import Foundation
-import UIKit
-
-enum ImageSource: Hashable {
-    case image(UIImage)
-    case imageURL(URL)
-    var isLocal: Bool {
-        switch self {
-            case .image: return true
-            case .imageURL: return false
-        }
-    }
-}
-
-protocol ImageLoaderProtocol {
-    func loadImage(from url: URL, completion: @escaping (Result<UIImage, Error>) -> Void)
-}
-
-struct DefaultImageLoader: ImageLoaderProtocol {
-    enum ImageError: Error {
-        case unknown
-        case corruptedData
-    }
-    
-    init() {}
-    
-    func loadImage(from url: URL, completion: @escaping (Result<UIImage, Error>) -> Void) {
-        let request = URLRequest(url: url, cachePolicy: .reloadIgnoringCacheData)
-        let session = URLSession(configuration: URLSessionConfiguration.default, delegate: nil, delegateQueue: nil)
-        let sessionDataTask = session.dataTask(with: request, completionHandler: { (data: Data?, _: URLResponse?, error: Error?) in
-            DispatchQueue.global(qos: .utility).async {
-                guard let imageData = data else {
-                    DispatchQueue.main.async {
-                        guard let error else {
-                            completion(.failure(ImageError.unknown))
-                            return
-                        }
-                        completion(.failure(error))
-                    }
-                    return
-                }
-                guard let image = UIImage(data: imageData) else {
-                    DispatchQueue.main.async {
-                        completion(.failure(ImageError.corruptedData))
-                    }
-                    return
-                }
-                DispatchQueue.main.async {
-                    completion(.success(image))
-                }
-            }
-        })
-        sessionDataTask.resume()
-    }
-}
